upgrade to molecule 2.x; fix linter; update contributor guidelines; upgrade testinfra tests
