+++ .yamllint
@@ -1,0 +1,14 @@
+extends: default
+ignore: |
+  .travis/
+  .travis.yml
+  meta/
+
+rules:
+  braces:
+    max-spaces-inside: 1
+    level: error
+  brackets:
+    max-spaces-inside: 1
+    level: error
+  line-length: disable
