+++ LICENSE
@@ -1,6 +1,6 @@
 The MIT License (MIT)
 
-Copyright (c) 2017-2018 Pawel Krupa and Roman Demachkovych
+Copyright (c) 2017-2018 Pawel Krupa, Roman Demachkovych
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
