+++ defaults/main.yml
@@ -12,19 +12,19 @@ alertmanager_web_external_url: 'http://localhost:9093/'
 alertmanager_resolve_timeout: 3m
 
 alertmanager_config_flags_extra: {}
-#alertmanager_config_flags_extra:
-#  data.retention: 10
+# alertmanager_config_flags_extra:
+#   data.retention: 10
 
 # SMTP default params
 alertmanager_smtp: {}
-#alertmanager_smtp:
-#  from: ''
-#  smarthost: ''
-#  auth_username: ''
-#  auth_password: ''
-#  auth_secret: ''
-#  auth_identity: ''
-#  auth_require_tls: "True"
+# alertmanager_smtp:
+#   from: ''
+#   smarthost: ''
+#   auth_username: ''
+#   auth_password: ''
+#   auth_secret: ''
+#   auth_identity: ''
+#   auth_require_tls: "True"
 
 # Default values you can see here -> https://prometheus.io/docs/alerting/configuration/
 alertmanager_slack_api_url: ''
@@ -35,64 +35,64 @@ alertmanager_hipchat_auth_token: ''
 
 # First read: https://github.com/prometheus/alertmanager#high-availability
 alertmanager_mesh: {}
-#alertmanager_mesh:
-#  peer-id: "{{ ansible_default_ipv4.macaddress }}"
-#  listen-address: "{{ ansible_default_ipv4.address }}:6783"
-#  nickname: "{{ ansible_hostname }}"
-#  peers:
-#    - "{{ ansible_default_ipv4.address }}:6783"
-#    - "demo.cloudalchemy.org:6783"
+# alertmanager_mesh:
+#   peer-id: "{{ ansible_default_ipv4.macaddress }}"
+#   listen-address: "{{ ansible_default_ipv4.address }}:6783"
+#   nickname: "{{ ansible_hostname }}"
+#   peers:
+#     - "{{ ansible_default_ipv4.address }}:6783"
+#     - "demo.cloudalchemy.org:6783"
 
 alertmanager_receivers: []
-#alertmanager_receivers:
-#- name: slack
-#  slack_configs:
-#  - send_resolved: true
-#    api_url: $slack_api_url
-#    channel: '#alerts'
+# alertmanager_receivers:
+#   - name: slack
+#     slack_configs:
+#       - send_resolved: true
+#         api_url: $slack_api_url
+#         channel: '#alerts'
 
 alertmanager_inhibit_rules: []
 
 alertmanager_route: {}
-#alertmanager_route:
-#  group_by: ['alertname', 'cluster', 'service']
-#  group_wait: 30s
-#  group_interval: 5m
-#  repeat_interval: 4h
-#  receiver: slack
+# alertmanager_route:
+#   group_by: ['alertname', 'cluster', 'service']
+#   group_wait: 30s
+#   group_interval: 5m
+#   repeat_interval: 4h
+#   receiver: slack
 
 alertmanager_child_routes: []
-#alertmanager_child_routes:
-#  # This routes performs a regular expression match on alert labels to
-#  # catch alerts that are related to a list of services.
-#  - match_re:
-#      service: ^(foo1|foo2|baz)$
-#    receiver: team-X-mails
-#    # The service has a sub-route for critical alerts, any alerts
-#    # that do not match, i.e. severity != critical, fall-back to the
-#    # parent node and are sent to 'team-X-mails'
-#    routes:
-#    - match:
-#        severity: critical
-#      receiver: team-X-pager
-#  - match:
-#      service: files
-#    receiver: team-Y-mails
-#    routes:
-#    - match:
-#        severity: critical
-#      receiver: team-Y-pager
-#  # This route handles all alerts coming from a database service. If there's
-#  # no team to handle it, it defaults to the DB team.
-#  - match:
-#      service: database
-#    receiver: team-DB-pager
-#    # Also group alerts by affected database.
-#    group_by: [alertname, cluster, database]
-#    routes:
-#    - match:
-#        owner: team-X
-#      receiver: team-X-pager
-#    - match:
-#        owner: team-Y
-#      receiver: team-Y-pager
+# alertmanager_child_routes:
+#   # This routes performs a regular expression match on alert labels to
+#   # catch alerts that are related to a list of services.
+#     - match_re:
+#         service: ^(foo1|foo2|baz)$
+#       receiver: team-X-mails
+#       # The service has a sub-route for critical alerts, any alerts
+#       # that do not match, i.e. severity != critical, fall-back to the
+#       # parent node and are sent to 'team-X-mails'
+#       routes:
+#         - match:
+#             severity: critical
+#           receiver: team-X-pager
+#     - match:
+#         service: files
+#       receiver: team-Y-mails
+#       routes:
+#         - match:
+#             severity: critical
+#           receiver: team-Y-pager
+#     # This route handles all alerts coming from a database service. If there's
+#     # no team to handle it, it defaults to the DB team.
+#     - match:
+#         service: database
+#       receiver: team-DB-pager
+#       # Also group alerts by affected database.
+#       group_by: [alertname, cluster, database]
+#       routes:
+#         - match:
+#             owner: team-X
+#           receiver: team-X-pager
+#         - match:
+#             owner: team-Y
+#           receiver: team-Y-pager
