+++ handlers/main.yml
@@ -2,7 +2,7 @@
 - name: restart alertmanager
   become: true
   systemd:
-    daemon_reload: yes
+    daemon_reload: true
     name: alertmanager
     state: restarted
