+++ molecule/alternative/tests/test_alternative.py
@@ -1,0 +1,43 @@
+import pytest
+import os
+import testinfra.utils.ansible_runner
+
+testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
+
+
+@pytest.mark.parametrize("dirs", [
+    "/etc/alertmanager",
+    "/etc/alertmanager/templates",
+    "/var/lib/alertmanager"
+])
+def test_directories(host, dirs):
+    d = host.file(dirs)
+    assert d.is_directory
+    assert d.exists
+
+
+@pytest.mark.parametrize("files", [
+    "/usr/local/bin/alertmanager",
+    "/usr/local/bin/amtool",
+    "/etc/alertmanager/alertmanager.yml",
+    "/etc/systemd/system/alertmanager.service"
+])
+def test_files(host, files):
+    f = host.file(files)
+    assert f.exists
+    assert f.is_file
+
+
+def test_service(host):
+    s = host.service("alertmanager")
+    assert s.is_enabled
+    assert s.is_running
+
+
+@pytest.mark.parametrize("sockets", [
+    "tcp://127.0.0.1:9093",
+    "tcp://127.0.0.1:6783"
+])
+def test_socket(host, sockets):
+    assert host.socket(sockets).is_listening
