+++ molecule/default/playbook.yml
@@ -1,0 +1,5 @@
+---
+- hosts: all
+  any_errors_fatal: true
+  roles:
+    - ansible-alertmanager
