--- tests/vars.yml
@@ -1,21 +1,0 @@
-alertmanager_slack_api_url: "http://example.com"
-alertmanager_listen_address: "127.0.0.1:9093"
-alertmanager_receivers:
-- name: slack
-  slack_configs:
-  - send_resolved: true
-    api_url: $slack_api_url
-    channel: '#alerts'
-alertmanager_route:
-  group_by: ['alertname', 'cluster', 'service']
-  group_wait: 30s
-  group_interval: 5m
-  repeat_interval: 3h
-  receiver: slack
-alertmanager_mesh:
-  peer-id: "{{ ansible_default_ipv4.macaddress }}"
-  listen-address: "127.0.0.1:6783"
-  nickname: "{{ ansible_hostname }}"
-  peers:
-    - "127.0.0.1:6783"
-    - "demo.cloudalchemy.org:6783"
