+++ vars/main.yml
@@ -1,3 +1,4 @@
+---
 go_arch_map:
   i386: 386
   x86_64: amd64
